Edit the result files and scripts
In this commit, the detailed results of model estimation such as the .iqtree, .log were included, the results folder was reconstructed to make it more clear.

Furthermore, the structures of the results of step 2 and 3 were also adjusted to make it more clear.

Furthermore, more scripts used in the analysis were added.

Finally, corresponding naming of the files of avian and ascomycota datasets were changed into bird and sacfungi